Reformat functions to fix auto-whitespace
For some reason, if a function is defined like this:

    function svndiff() {
    blah blah blah
    }

vim will not automatically indent the body.  However, if the function is
instead defined like this:

    svndiff() {
            blah blah blah
    }

vim will indent the body of the function one shiftwidth.  So I am going
to do that since it makes the function bodies clearer.